Switch dart internal to an http endpoint
