+++ app_dart/lib/cocoon_service.dart
@@ -6,7 +6,7 @@ export 'src/foundation/utils.dart';
 export 'src/model/appengine/service_account_info.dart';
 export 'src/request_handlers/check_flaky_builders.dart';
 export 'src/request_handlers/create_branch.dart';
-export 'src/request_handlers/dart_internal_subscription.dart';
+export 'src/request_handlers/dart_internal_task_result.dart';
 export 'src/request_handlers/file_flaky_issue_and_pr.dart';
 export 'src/request_handlers/flush_cache.dart';
 export 'src/request_handlers/get_authentication_status.dart';
